Change indexing for identifying no key
Using the single bracket [] treated the vector as a list, and evaluated
it as a data.frame or tbl.  Thus, when using is.numeric always returned
FALSE.  The double bracket [[]] evaluates the specific element, thus
treating the vector as numeric, character, integer, etc... which allows
for TRUE to be returned.